[AWIBOF-8269] Split command_processor.cpp into multiple files
Split command_processor.cpp into multiple files.

Previously, the file was too long so the readability
was poor. It is difficult to split it into
multiple classes also.

This commit splits the .cpp file only (not class itself),
into several command categories.
